Added files from bananasrlowkeygood to FTK for long-term maintenance.
Previously hosting a few files on my personal site to avoid having to create a new Zenodo, now that I have some free time, I realize that this is not the best solution for long term maintanance. For now, will park them under tests>data>ftk-tutorial and reupdate the WIKI once this PR is approved.